perf: add getProgramBySlug with display-only key status
- extracts applyKeyStatusForDisplay for cron-gap ui updates.
- keeps getProgramWithUpdatedKeys as alias to getProgramBySlug.
